Spring security with JWT tokens added

+++ src/main/java/user_service/service/CardService.java
@@ -17,6 +17,7 @@ import user_service.entity.User;
 import user_service.exception.CardNotFoundException;
 import user_service.exception.CardNumberNotUniqueException;
 import user_service.exception.CardsNotFoundException;
+import user_service.exception.InvalidCardOwnerException;
 import user_service.exception.UserNotFoundException;
 import user_service.mapper.CardMapper;
 
@@ -38,26 +39,31 @@ public class CardService {
         }
     }
 
-    private User getCardOwnerUserById(CardRequestDto cardRequestDto) throws UserNotFoundException {
-        final long userId = cardRequestDto.getUserId();
-        return userDao.findUserById(userId)
-                .orElseThrow(() -> new UserNotFoundException(userId));
+    private User getCurrentUser(Long userId) {
+        return userDao.findUserById(userId).orElseThrow(() -> new UserNotFoundException(userId));
+    }
+
+    private boolean isAuthenticatedUserNotCardOwner(Card card, long currentUserId) {
+        return card.getUser().getId() != currentUserId;
     }
 
     @CachePut(value = "card:id", key = "#result.id")
-    @CacheEvict(value = "user:id", key = "#cardRequestDto.userId")
-    public CardResponseDto createCard(CardRequestDto cardRequestDto) {
+    @CacheEvict(value = "user:id", key = "#result.userId")
+    public CardResponseDto createCard(CardRequestDto cardRequestDto, Long userId) {
         validateCardNumberUnique(cardRequestDto);
         Card card = cardMapper.toCard(cardRequestDto);
-        card.setUser(getCardOwnerUserById(cardRequestDto));
+        card.setUser(getCurrentUser(userId));
         card = cardDao.save(card);
         return cardMapper.toResponseDto(card);
     }
 
     @Cacheable("card:id")
-    public CardResponseDto getCardById(Long id) {
+    public CardResponseDto getCardById(Long id, Long userId) {
         Card card = cardDao.findCardById(id)
                 .orElseThrow(() -> new CardNotFoundException(id));
+        if (isAuthenticatedUserNotCardOwner(card, userId)) {
+            throw new InvalidCardOwnerException("The card is owned by another user");
+        }
         return cardMapper.toResponseDto(card);
     }
 
@@ -65,9 +71,9 @@ public class CardService {
     @Caching(put = {
             @CachePut(value = "card:id", key = "#id")
     }, evict = {
-            @CacheEvict(value = "user:id", key = "#cardRequestDto.userId")
+            @CacheEvict(value = "user:id", key = "#result.userId")
     })
-    public CardResponseDto updateCard(CardRequestDto cardRequestDto, long id) {
+    public CardResponseDto updateCard(CardRequestDto cardRequestDto, long id, Long userId) {
         Card card = cardDao.findCardById(id)
                 .orElseThrow(() -> new CardNotFoundException(id));
         if (!cardRequestDto.getNumber().equals(card.getNumber())) {
@@ -75,17 +81,15 @@ public class CardService {
         }
 
         cardMapper.updateCardFromDto(cardRequestDto, card);
-        long previousUserId = card.getUser().getId();
-        if (previousUserId != cardRequestDto.getUserId()) {
-            cacheManager.getCache("user:id").evict(previousUserId);
-            card.setUser(getCardOwnerUserById(cardRequestDto));
+        if (isAuthenticatedUserNotCardOwner(card, userId)) {
+            throw new InvalidCardOwnerException("The card is owned by another user");
         }
 
         cardDao.save(card);
         return cardMapper.toResponseDto(card);
     }
 
-    public List<CardResponseDto> getCardsByIds(List<Long> ids) {
+    public List<CardResponseDto> getCardsByIds(List<Long> ids, Long currentUserId) {
         List<Card> cards = cardDao.findCardsByIdIn(ids);
         if (cards == null || cards.isEmpty()) {
             throw new CardsNotFoundException();
@@ -93,15 +97,23 @@ public class CardService {
         return cards
                 .stream()
                 .map(cardMapper::toResponseDto)
+                .peek(card -> {
+                    if (!card.getUserId().equals(currentUserId)) {
+                        throw new InvalidCardOwnerException("The card is owned by another user");
+                    }
+                })
                 .collect(Collectors.toList());
     }
 
     @Transactional
     @CacheEvict(value = "card:id", key = "#id")
-    public void deleteCard(Long id) {
+    public void deleteCard(Long id, Long userId) {
         Card card = cardDao.findCardById(id)
                 .orElseThrow(() -> new CardNotFoundException(id));
 
+        if (isAuthenticatedUserNotCardOwner(card, userId)) {
+            throw new InvalidCardOwnerException("The card is owned by another user");
+        }
         cacheManager.getCache("user:id").evict(card.getUser().getId());
         cardDao.delete(card);
     }
